xGroup: Ensure group members are always returned as an array
PowerShell will unpack single element arrays on function return by
default. That's problematic when we explicitly expect an array.

Note that prefixing the return operation in Get-MembersOnFullSKU and
Get-MembersOnNanoServer with ',' to indicate the array should not be
unrolled introduces separate issues as the underlying mechanism will
wrap the "inner" array in an "outer" array, the latter being unpacked
automatically when passed down the pipeline. That means:

- Pester tests will fail when no group members are expected as it will
  encounter an empty array (the "inner" array) instead of $null.
- Pester tests will fail on testing expected group members as it will
  encounter an array (the "inner" array) instead of the enumerated
  members from the result passed down the pipeline.

Both of the above are known Pester issues (see Pester Issue #1154).

We could update the tests, but behind the scenes Pester is complaining
about a change in returned data types, and that feels problematic. So
instead check if the returned data is a string (indicating a single
group member) and convert it to an object array if so.