xGroup: Add additional unit tests to verify single member behaviour
This introduces two new tests which verify the expected behaviour on
returning the members of a group which has a single member (a single
element array). One test is for Nano server and the other for the full
SKU as they have different codepaths. The existing two group member
tests are renamed to explicitly reflect they're checking behaviour on
returning a group with multiple members.